+++ point3D.hpp
@@ -11,7 +11,7 @@ class Point3D
 	public:
 		Point3D() = default;
 		Point3D(Point3D const& p);
-		Point3D(int xx, int yy, int zz);
+		Point3D(double xx, double yy, double zz);
 		~Point3D();
 		double getNorm() const { return sqrt(pow(_x,2)+pow(_y,2)+pow(_z,2)); }
 		double getX() const {return _x;};
